requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==1.1.1
django-stdimage==5.1.1
gunicorn==20.0.4
Pillow==7.1.1
progressbar2==3.50.1
psycopg2-binary==2.8.5
PyMySQL==0.9.3
python-utils==2.4.0
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
whitenoise==5.0.1
